requirements.txt
django==1.11.15
django-cors-headers==2.4.0
django-cors-middleware==1.3.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
pandas==0.21.0
python-docx==0.8.7
pytest==3.5.1
natsort==5.5.0
docx2txt==0.7
ReadabilityCalculator==0.2.37
PyMySQL==0.9.3
textstat==0.5.6
nltk==3.2.1
numpy==1.15.1
glob2==0.6
spacy==2.0.13
language-check==1.1
scikit-learn==0.19.1
scipy==1.0.0
xlsxwriter==1.1.8
xlrd==1.2.0
